requirements.txt
# Core configuration / validation
pydantic<2

# Tokenization & embeddings
tiktoken>=0.7
sentence-transformers>=2.2.2

# Vector search
faiss-cpu>=1.12.0

# Numerical & data frames (need wheels for 3.14)
numpy>=2.0
pandas>=2.2.2

# Utilities
pyyaml>=6.0
tenacity>=9.0
loguru==0.7.0

# PDF & text extraction
pypdf>=5.0
PyMuPDF>=1.24

# Database / optional pgvector
psycopg2-binary>=2.9.6 ; platform_system != "Windows"

# OpenAI / Azure APIs (new OpenAI SDK; works with our embedder)
openai>=1.0.0,<3
azure-search-documents>=11.6.0
azure-core>=1.30.0

# OCR
pytesseract>=0.3.10
Pillow>=11.0.0
